Refactor SelectExpenseCommandTest and SelectExpenseCommandParserTest
