+++ src/test/java/seedu/address/logic/parser/expenseparsers/SelectExpenseCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.expenseparsers;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,8 +8,6 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_EXPENSE;
 import org.junit.Test;
 
 import seedu.address.logic.commands.expensecommands.SelectExpenseCommand;
-import seedu.address.logic.parser.expenseparsers.DeleteExpenseCommandParserTest;
-import seedu.address.logic.parser.expenseparsers.SelectExpenseCommandParser;
 
 /**
  * Test scope: similar to {@code DeleteExpenseCommandParserTest}.
